+++ lib/rley/parser/chart.rb
@@ -7,10 +7,19 @@ module Rley # This module is used as a namespace
     # A one-dimensional array with n + 1 entries (n = number of input tokens).
     class Chart
       attr_reader(:state_sets)
+      
+      # The level of trace details reported on stdout during the parse.
+      # The possible values are:
+      # 0: No trace output (default case)
+      # 1: Show trace of scanning and completion rules
+      # 2: Same as of 1 with the addition of the prediction rules
+      attr_reader(:tracer)
 
-      def initialize(startDottedItem, tokenCount)
+      # @param aTracerLevel [ParseTracer] A tracer object. 
+      def initialize(startDottedItem, tokenCount, aTracer)
+        @tracer = aTracer
         @state_sets = Array.new(tokenCount + 1) { |_| StateSet.new }
-        push_state(startDottedItem, 0, 0)
+        push_state(startDottedItem, 0, 0, :start_rule)
       end
 
       # The dotted item/rule used to seed the parse chart.
@@ -38,9 +47,23 @@ module Rley # This module is used as a namespace
       end
 
       # Push a parse state for the chart entry with given index
-      def push_state(aDottedItem, anOrigin, anIndex)
+      def push_state(aDottedItem, anOrigin, anIndex, aReason)
         new_state = ParseState.new(aDottedItem, anOrigin)
-        self[anIndex].push_state(new_state)
+        pushed = self[anIndex].push_state(new_state)
+        if pushed && tracer.level > 0
+          case aReason
+            when :start_rule, :prediction
+              tracer.trace_prediction(anIndex, new_state)
+              
+            when :scanning
+               tracer.trace_scanning(anIndex, new_state)
+               
+            when :completion
+               tracer.trace_completion(anIndex, new_state)
+            else
+              raise NotImplementedError, "Unknown push_state mode #{aReason}"
+          end
+        end
       end
     end # class
   end # module